move snackbarAleret file
